pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "linkedin_copilot"
version = "0.1.0"
description = "Local-first LinkedIn job search and application copilot using Browser Use and Ollama."
readme = "linkedin_copilot/README.md"
requires-python = ">=3.11"
authors = [
  { name = "Your Name" }
]
license = { text = "MIT" }
dependencies = [
  "pydantic>=2.9",
  "pydantic-settings>=2.6",
  "typer>=0.12",
  "rich>=13.9",
  "loguru>=0.7",
  "python-dotenv>=1.0",
  "pyyaml>=6.0",
  "browser-use>=0.1.16",
  "playwright>=1.49",
  "langchain-core>=0.3.13",
  "langchain-ollama>=0.2.0",
  "aiohttp>=3.10",
  "tenacity>=9.0",
  "pandas>=2.2",
  "fastapi>=0.115",
  "uvicorn[standard]>=0.32",
  "jinja2>=3.1",
  "python-multipart>=0.0.9",
  "tavily-python",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3",
  "pytest-asyncio>=0.24",
  "mypy>=1.12",
  "ruff>=0.8",
]

[project.scripts]
linkedin_copilot = "linkedin_copilot.cli:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true

[tool.ruff]
line-length = 100
target-version = "py311"

